fix: page loading
